Reduce Meme Lab card duplication
